+++ mvc/src/main/java/nextstep/mvc/controller/scanner/ControllerScanner.java
@@ -1,6 +1,11 @@
 package nextstep.mvc.controller.scanner;
 
+import static java.util.stream.Collectors.toMap;
+
+import java.lang.reflect.InvocationTargetException;
+import java.util.Map;
 import java.util.Set;
+import java.util.function.Function;
 import nextstep.web.annotation.Controller;
 import org.reflections.Reflections;
 
@@ -15,8 +20,19 @@ public class ControllerScanner {
         return INSTANCE;
     }
 
-    public Set<Class<?>> getAllAnnotations(Object... basePackage) {
+    public Map<Class<?>, Object> getControllers(final Object... basePackage) {
         Reflections reflections = new Reflections(basePackage);
-        return reflections.getTypesAnnotatedWith(Controller.class);
+        Set<Class<?>> mappedControllers = reflections.getTypesAnnotatedWith(Controller.class);
+
+        return mappedControllers.stream()
+            .collect(toMap(Function.identity(), this::getInitializedController));
+    }
+
+    private Object getInitializedController(final Class<?> mappedController) {
+        try {
+            return mappedController.getConstructor().newInstance();
+        } catch (NoSuchMethodException | InvocationTargetException | InstantiationException | IllegalAccessException e) {
+            throw new IllegalArgumentException(e);
+        }
     }
 }
