+++ mvc/src/test/java/nextstep/mvc/controller/scanner/ControllerScannerTest.java
@@ -4,9 +4,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.util.List;
-import java.util.Set;
+import java.util.Map;
 import java.util.stream.Collectors;
-import nextstep.web.annotation.Controller;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,25 +15,16 @@ class ControllerScannerTest {
     @DisplayName("Controller Scanner는 BasePackage를 기준으로 @Controller가 사용된 모든 클래스를 찾는다.")
     void findAllClassesOfAnnotationController() {
         // given & when
-        Set<Class<?>> classes = ControllerScanner.getInstance().getAllAnnotations(
-            "nextstep.mvc.controller.scanner");
+        Map<Class<?>, Object> controllers = ControllerScanner.getInstance()
+            .getControllers("samples");
 
         // then
-        List<String> classNames = classes.stream()
+        List<String> extract = controllers.keySet().stream()
             .map(Class::getSimpleName)
             .collect(Collectors.toList());
 
         assertAll(
-            () -> assertThat(classNames).contains("AnnotationController")
+            () -> assertThat(extract).contains("TestController")
         );
-    }
-
-    @Controller
-    private class AnnotationController {
-
-    }
-
-    private class NotAnnotationController {
-
     }
 }
